requirements.txt
streamlit==1.36.0
numpy==1.26.4
scipy==1.11.4
matplotlib==3.8.2
pillow==10.2.0
tensorflow==2.15.0
typing_extensions==4.9.0